bmc-nix: Switch to usize #BDK-356
- prefer usize for generation numbers (counts/indices)
- keep u32 only for JSON schema contract types (PackageIndex, FactoryServerEntry)

+++ bmc-nix/src/types.rs
@@ -146,7 +146,7 @@ pub struct ManifestPackage {
 /// Profile generation metadata
 #[derive(Debug, Clone)]
 pub struct ProfileGeneration {
-    pub number: u32,
+    pub number: usize,
     pub path: PathBuf,
     pub manifest: Manifest,
 }
